brain_tumor_classifier/requirements.txt
tensorflow==2.13.0
keras==2.13.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.0
openpyxl==3.1.2
streamlit==1.28.0
